🗑️ Removed all instances of sentry.hamravesh.com
In this update, the domain `sentry.hamravesh.com` has been removed from multiple YAML files within the repository. This change affects the following files: `GoodbyeAds.yaml`, `category-ads-all-2.yaml`, `clash_rules_ads-persian-blocker.yaml`, and `add.yaml`.

The removal of this domain was executed to ensure the ad-blocking rules are up-to-date and to prevent any potential issues related to outdated or unnecessary entries. Each instance of this domain was carefully identified and eliminated to maintain the integrity and efficiency of the ad-blocking configurations.

Make sure to review the changes in the `/home/taha/WebstormProjects/clash-rules` directory to verify that the modifications align with your project's requirements.